feat: refactor to ECMAScript
