+++ src/modules/esm.mjs
@@ -1,0 +1,37 @@
+import path, { sep, dirname } from 'path';
+import { release, version } from 'os';
+import { createServer as createServerHttp } from 'http';
+import { fileURLToPath } from 'url';
+import { promises as fs } from 'fs';
+import './files/c.js';
+
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+const randomValue = Math.random();
+
+const jsonFilePath = path.join(__dirname, 'files', `${randomValue > 0.5 ? 'a' : 'b'}.json`);
+const fileContent = await fs.readFile(jsonFilePath, 'utf-8');
+const unknownObject = JSON.parse(fileContent);
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+    console.log(`Server is listening on port ${PORT}`);
+    console.log('To terminate it, use Ctrl+C combination');
+});
+
+export { unknownObject, myServer };
